Created a separate .jsx file for components on product page
